backend/requirements.txt
flask
python-dotenv
psycopg2-binary
bcrypt
ollama
flask-cors